# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Tecni Motors",
        "description": "API's for TecniMotors",
        "version": "1.0"
    },
    "paths": {
...
        "/pedidos-backend/public/api/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "description": "Create a new user",
                "operationId": "ee10365648d90eb967aa44f2f107f238",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return user created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TypeUser": {
                "title": "TypeUser",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Administrador"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "username",
                    "password",
                    "person_id",
                    "typeofUser_id"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "person_id": {
                        "description": "Foreign key referencing person",
                        "type": "number",
                        "example": "1"
                    },
                    "typeofUser_id": {
                        "description": "Foreign key referencing TypeOfUser",
                        "type": "number",
                        "example": "1"
                    },
                    "TypeUser": {
                        "$ref": "#/components/schemas/TypeUser"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "title": "UserRequest",
                "description": "User request model",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "Username"
                    },
                    "password": {
                        "type": "string",
                        "example": "Password"
                    },
                    "names": {
                        "type": "string",
                        "example": "names"
                    },
                    "motherSurname": {
                        "type": "string",
                        "example": "motherSurname"
                    },
                    "fatherSurname": {
                        "type": "string",
                        "example": "fatherSurname"
                    },
                    "email": {
                        "type": "string",
                        "example": "email"
                    },
                    "address": {
                        "type": "string",
                        "example": "address"
                    },
                    "phone": {
                        "type": "string",
                        "example": "phone"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}